services/api_gateway/requirements.txt
fastapi
uvicorn
pydantic
requests
redis[asyncio]
python-dotenv
gunicorn
prometheus-client
